# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for the ECSite",
        "title": "ECSite API",
        "contact": {},
        "license": {
            "name": "Techforward Inc.",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:1323",
    "basePath": "/",
    "paths": {
...
        "/item": {
            "get": {
                "description": "商品をリストで返す",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Item"
                ],
                "summary": "商品リスト",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/models.Item"
                        }
                    },
                    "404": {}
                }
            },
            "post": {
                "description": "create new item in a ECsite",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Item"
                ],
                "summary": "create new item",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/models.Item"
                        }
                    }
                }
            }
        },
        "/item/{id}": {
            "get": {
                "description": "商品を返す",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Item"
                ],
                "summary": "商品",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/models.Item"
                        }
                    },
                    "404": {}
                }
            },
            "put": {
                "description": "updating item in a ECsite",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Item"
                ],
                "summary": "Update item",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/models.Item"
                        }
                    }
                }
            },
            "delete": {
                "description": "deleting item in a ECsite",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Item"
                ],
                "summary": "Delete item",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/order": {
            "get": {
                "description": "商品をリストで返す",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order"
                ],
                "summary": "商品リスト",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/models.Order"
                        }
                    },
                    "404": {}
                }
            },
            "post": {
                "description": "create new order in a ECsite",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order"
                ],
                "summary": "create new order",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/models.Order"
                        }
                    }
                }
            }
        },
        "/order/{id}": {
            "get": {
                "description": "商品を返す",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order"
                ],
                "summary": "商品",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/models.Order"
                        }
                    },
                    "404": {}
                }
            },
            "put": {
                "description": "updating order in a ECsite",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order"
                ],
                "summary": "Update order",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/models.Order"
                        }
                    }
                }
            },
            "delete": {
                "description": "deleting order in a ECsite",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order"
                ],
                "summary": "Delete order",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/payment/stripe": {
            "post": {
                "description": "決済を行う",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Stripe"
                ],
                "summary": "決済",
                "responses": {
                    "200": {},
                    "404": {}
                }
            }
        },
        "/user": {
            "get": {
                "description": "商品をリストで返す",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "商品リスト",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/models.User"
                        }
                    },
                    "404": {}
                }
            },
            "post": {
                "description": "create new user in a ECsite",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "create new user",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/models.User"
                        }
                    }
                }
            }
        },
        "/user/{id}": {
            "get": {
                "description": "商品を返す",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "商品",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/models.User"
                        }
                    },
                    "404": {}
                }
            },
            "put": {
                "description": "updating user in a ECsite",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Update user",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/models.User"
                        }
                    }
                }
            },
            "delete": {
                "description": "deleting user in a ECsite",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Delete user",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Item": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer",
                    "format": "int64",
                    "example": 4
                },
                "context": {
                    "type": "string",
                    "example": "item context"
                },
                "created_at": {
                    "description": "CreatedAt, UpdatedAt, DeleteAt という time.Time 型のメンバを追加すると、\nGORM側でここに追加・更新・削除日時を自動的に入るようになる。",
                    "type": "string",
                    "example": "2019-06-14 05:59:41"
                },
                "name": {
                    "type": "string",
                    "example": "item name"
                },
                "priority": {
                    "type": "integer",
                    "format": "int64",
                    "example": 1
                },
                "ulid": {
                    "type": "string",
                    "example": "0000XSNJG0MQJHBF4QX1EFD6Y3"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2019-06-15 15:01:52"
                }
            }
        },
...
        "models.Order": {
            "type": "object",
            "properties": {
                "addressUlid": {
                    "type": "string",
                    "example": "0000XSNJG0JFGRHF4QX1EFDAH2"
                },
                "created_at": {
                    "type": "string",
                    "example": "2019-06-14 05:59:41"
                },
                "isCanceled": {
                    "type": "integer",
                    "example": 0
                },
                "status": {
                    "type": "integer",
                    "example": 0
                },
                "ulid": {
                    "type": "string",
                    "example": "0000XSNJG0MQJHBF4QX1EFD6Y3"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2019-06-15 15:01:52"
                },
                "userUlid": {
                    "type": "string",
                    "example": "0000XSNJG0JFGRHF4QX1EFD6Y3"
                }
            }
        },
        "models.User": {
            "type": "object",
            "properties": {
                "addressUlid": {
                    "type": "string",
                    "example": "0000XSHJW0MQJ3JK3QX1EFD6Y3"
                },
                "cartUlid": {
                    "type": "string",
                    "example": "0000XSHJW0MQJHBF4QX1EFD6Y3"
                },
                "created_at": {
                    "type": "string",
                    "example": "2019-06-14 05:59:41"
                },
                "email": {
                    "type": "string",
                    "example": "user@gmail.com"
                },
                "firebaseUid": {
                    "type": "string",
                    "example": "0gwEMgeG0ePdEZOZV6ZchBgWyE42"
                },
                "name": {
                    "type": "string",
                    "example": "user124"
                },
                "stripeId": {
                    "type": "string",
                    "example": "cus_fherfjwh234rhej"
                },
                "ulid": {
                    "description": "gorm:\"primary_key\"を消すとテーブルが空っぽになる",
                    "type": "string",
                    "example": "0000XSNJG0MQJHBF4QX1EFD6Y3"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2019-06-15 15:01:52"
                }
            }
        },
...
    }
}